+++ test/FoldingStrats/FoldingStratTestCase.ts
@@ -1,0 +1,12 @@
+type FoldingStratTestCase = {
+   name: string,
+   tokenAddress: string;
+   strategyAddress: string;
+   snowglobeAddress: string;
+   slot: number;
+   amount: string;
+   fold: boolean;
+   controller: string;
+};
+
+export type { FoldingStratTestCase };
