+++ test/LPStrats/LPStratTestCase.ts
@@ -1,0 +1,6 @@
+type LPStratTestCase = {
+   name: string;
+   controllerAddress: string; 
+   snowglobeAddress: string; 
+}
+export type { LPStratTestCase };
