+++ test/Zapper/joe-zapper-tests.ts
@@ -4,7 +4,7 @@
  * 
  *  ***/
 
-const { doZapperTests } = require("../zapper-test");
+import { doZapperTests } from "./zapper-test";
 
 const tests = [
     {
@@ -22,8 +22,14 @@ const tests = [
 ];
 
 
+describe("Joe Zapper Tests", function ()  {
 for (const test of tests) {
-    describe(test.name, async () => {
-        doZapperTests(test.name, test.snowglobeAddress, "TraderJoe", test.gaugeAddress, test.controller);
-    });
-}
+        doZapperTests(
+           test.name, 
+           test.snowglobeAddress, 
+           "TraderJoe", 
+           test.gaugeAddress, 
+           test.controller
+        );
+   }
+});
