+++ test/utils/tokens.ts
@@ -1,0 +1,6 @@
+export const TOKEN_ADDRS = {
+   "WAVAX": "",
+}
+
+export const CONTRACT_ADDS = {
+}
